Rename mtcnn file
